+++ src/ExchangeRateProvider.php
@@ -16,9 +16,9 @@ interface ExchangeRateProvider
      * @param string $sourceCurrencyCode The source currency code.
      * @param string $targetCurrencyCode The target currency code.
      *
-     * @return BigNumber|int|float|string The exchange rate.
+     * @return BigNumber|int|string The exchange rate.
      *
      * @throws CurrencyConversionException If the exchange rate is not available.
      */
-    public function getExchangeRate(string $sourceCurrencyCode, string $targetCurrencyCode): BigNumber|int|float|string;
+    public function getExchangeRate(string $sourceCurrencyCode, string $targetCurrencyCode): BigNumber|int|string;
 }